Issue #39: Rebrand Codebase from Sentry Software to MetricsHub * Replace sentry software by metricshub in packages names, docs and all the project files
