+++ .github/workflows/build.yml
@@ -18,7 +18,7 @@ on:
         
 jobs:
   build:
-    uses: sentrysoftware/workflows/.github/workflows/maven-build.yml@v3
+    uses: metricshub/workflows/.github/workflows/maven-build.yml@v3
     with:
       jdkVersion: "17"
       debug: ${{ github.event_name == 'workflow_dispatch' && inputs.debug }}
