+++ .github/workflows/deploy.yml
@@ -10,7 +10,7 @@ on:
 
 jobs:
   deploy:
-    uses: sentrysoftware/workflows/.github/workflows/maven-central-deploy.yml@v3
+    uses: metricshub/workflows/.github/workflows/maven-central-deploy.yml@v3
     with:
       jdkVersion: "17"
     secrets: inherit
