+++ src/main/java/org/metricshub/xflat/exceptions/XFlatException.java
@@ -1,10 +1,10 @@
-package org.sentrysoftware.xflat.exceptions;
+package org.metricshub.xflat.exceptions;
 
 /*-
  * ╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲
  * XFlat Utility
  * ჻჻჻჻჻჻
- * Copyright 2023 Sentry Software
+ * Copyright (C) 2023 - 2025 MetricsHub
  * ჻჻჻჻჻჻
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,11 +20,15 @@ package org.sentrysoftware.xflat.exceptions;
  * ╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱╲╱
  */
 
-public class XFlatRunTimeException extends RuntimeException {
+public class XFlatException extends Exception {
 
 	private static final long serialVersionUID = 1L;
 
-	public XFlatRunTimeException(final String message) {
+	public XFlatException(final String message) {
 		super(message);
+	}
+
+	public XFlatException(final String message, final Exception e) {
+		super(message, e);
 	}
 }
